Fixed a bug where, when there is no Data folder, program crashes .

+++ ToNote/ViewModels/MainViewModel.cs
@@ -102,8 +102,10 @@
 
                     var todosFileName = "*_" + note.Name + "_TODO";
 
-                    var todoFiles = Directory.GetFiles("Data", todosFileName);
+                    if (!Directory.Exists("Data"))
+                        return;
 
+                    var todoFiles = Directory.GetFiles("Data", todosFileName);
                     foreach (var file in todoFiles)
                     {
                         if (File.Exists(file)) File.Delete(file);
